January Radarr upstream 11 (#904)
* Fixed: Treat TAoE and QxR as release groups

(cherry picked from commit radarr/radarr@8c94581cb6)

* New: Prioritize Exact Exception Release Group Matches over non-Exact Exception Groups

(cherry picked from commit 9906b95893660a929cb9fd1393e28f09d8ada354)

* Bump coverlet.collector to official 6.0.4

(cherry picked from commit radarr/radarr@eaeb668eb5)

* Avoid rewriting file names in builds

(cherry picked from commit gcp radarr/radarr@9959c658be)

* Fix code coverage on CI

(cherry picked from commit radarr/radarr@97c226c23c)

* Fixed: RlsGrp parser exeption for 126811

(cherry picked from commit radarr/radarr@64b2a10b3f)

* Improve error tracing in migrate app data folder

(cherry picked from commit radarr/radarr@554a54b009)

* Attempt to remove pid file only if config folder exists

(cherry picked from commit radarr/radarr@0297dba7f9)

* Bump Moq to 4.18.4

(cherry picked from commit radarr/radarr@52307038af)

* Switch HttpProxySettingsProviderFixture to test cases

(cherry picked from commit radarr/radarr@1e06fc5b43)

* Bump postcss to 8.5.6

(cherry picked from commit radarr/radarr@3c75250c08)

---------

+++ src/NzbDrone.Common/EnvironmentInfo/AppFolderFactory.cs
@@ -122,7 +122,7 @@ namespace NzbDrone.Common.EnvironmentInfo
             catch (Exception ex)
             {
                 _logger.Debug(ex, ex.Message);
-                throw new WhisparrStartupException("Unable to migrate DB from nzbdrone.db to {0}. Migrate manually", _appFolderInfo.GetDatabase());
+                throw new WhisparrStartupException(ex, "Unable to migrate DB from nzbdrone.db to {0}. Migrate manually", _appFolderInfo.GetDatabase());
             }
         }
 
@@ -199,7 +199,7 @@ namespace NzbDrone.Common.EnvironmentInfo
 
         private void RemovePidFile()
         {
-            if (OsInfo.IsNotWindows)
+            if (OsInfo.IsNotWindows && _diskProvider.FolderExists(_appFolderInfo.AppDataFolder))
             {
                 _diskProvider.DeleteFile(Path.Combine(_appFolderInfo.AppDataFolder, "whisparr.pid"));
             }
